4.24

+++ code/result.h
@@ -7,7 +7,7 @@ class Result {
 public:
 	int requestNum;
 	int receiveNum;
-	int income = 0;
+	double income = 0;
 	int cost = 0;
 	Graph topo;
 	RequestList requests;
